# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IMDb REST API",
        "version": "1.0",
        "description": "REST API to watch movies information"
    },
    "servers": [
        {
            "url": "https://imdb3.herokuapp.com/api/v1",
            "description": "Production"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development"
        }
    ],
    "paths": {
        "/users":{
            "post":{
                "description": "To add a new user. It sends a email to verify the email address and activate the user",
                "responses": {
                    "201":{
                        "description": "User added correctly"
                    }
                },
                "requestBody":{
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/Users"
                            }
                        }
                    }
                },
                "tags": ["Users"]
            },
            "get": {
                "description": "To get the data of the logged user",
                "responses": {
                    "200":{
                        "description": "User gotten correctly"
                    }
                },
                "security": [{
                    "jwt_token": []
                }],
                "tags": ["Users"]
            },
            "delete": {
                "description": "To delete the logged user",
                "responses": {
                    "200":{
                        "description": "User deleted"
                    }
                },
                "security": [{
                    "jwt_token": []
                }],
                "tags": ["Users"]
            },
            "put": {
                "description": "To update he current user",
                "responses": {
                    "200":{
                        "description": "User updated"
                    }
                },
                "requestBody":{
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/Users"
                            }
                        }
                    }
                },
                "security": [{
                    "jwt_token": []
                }],
                "tags": ["Users"]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt_token":{
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "Users":{
                "properties":{
                    "firstname":{
                        "type": "string",
                        "example": "Pablo"
                    },
                    "lastname":{
                        "type": "string",
                        "example": "Brown"
                    },
                    "email": {
                        "type": "string",
                        "example": "pablo@hotmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "pablo123"
                    }
                }
            },
...
        }
    }
}